Split project to two modules. Dialog module - for exchange messages with alice dialogs. Game module - for process darts game.
