+++ Dialog/src/main/java/com/kepler/model/request/Account_linking.java
@@ -1,0 +1,8 @@
+package com.kepler.model.request;
+
+import lombok.Data;
+
+@Data
+public class Account_linking {
+
+}
