+++ Dialog/src/main/java/com/kepler/model/request/Application.java
@@ -1,0 +1,8 @@
+package com.kepler.model.request;
+
+import lombok.Data;
+
+@Data
+public class Application {
+    private String application_id;
+}
