+++ Dialog/src/main/java/com/kepler/model/request/Entities.java
@@ -1,0 +1,10 @@
+package com.kepler.model.request;
+
+import lombok.Data;
+
+@Data
+public class Entities {
+    private String type;
+    private Tokens tokens;
+    private Object value;
+}
