+++ Dialog/src/main/java/com/kepler/model/request/InputData.java
@@ -1,0 +1,12 @@
+package com.kepler.model.request;
+
+import lombok.Data;
+
+@Data
+public class InputData {
+    private Meta meta;
+    private Session session;
+    private Request request;
+    private String version;
+}
+
