+++ Dialog/src/main/java/com/kepler/model/request/Interfaces.java
@@ -1,0 +1,10 @@
+package com.kepler.model.request;
+
+import lombok.Data;
+
+@Data
+public class Interfaces {
+    private Screen screen;
+    private Payments payments;
+    private Account_linking account_linking;
+}
