+++ Dialog/src/main/java/com/kepler/model/request/Markup.java
@@ -1,0 +1,8 @@
+package com.kepler.model.request;
+
+import lombok.Data;
+
+@Data
+public class Markup {
+    private Boolean dangerous_context;
+}
