+++ Dialog/src/main/java/com/kepler/model/request/Nlu.java
@@ -1,0 +1,10 @@
+package com.kepler.model.request;
+
+import lombok.Data;
+
+@Data
+public class Nlu {
+    private String[] tokens;
+    private Entities[] entities;
+    private Intents intents;
+}
