+++ Dialog/src/main/java/com/kepler/model/request/Request.java
@@ -1,0 +1,12 @@
+package com.kepler.model.request;
+
+import lombok.Data;
+
+@Data
+public class Request {
+    private String command;
+    private String original_utterance;
+    private Nlu nlu;
+    private Markup markup;
+    private String type;
+}
