+++ Dialog/src/main/java/com/kepler/model/request/Tokens.java
@@ -1,0 +1,9 @@
+package com.kepler.model.request;
+
+import lombok.Data;
+
+@Data
+public class Tokens {
+    private Integer start;
+    private Integer end;
+}
