+++ Dialog/src/main/java/com/kepler/model/response/Response.java
@@ -1,0 +1,10 @@
+package com.kepler.model.response;
+
+import lombok.Data;
+
+@Data
+public class Response {
+    private Boolean end_session;
+    private String text;
+    private String tts;
+}
