+++ Game/src/main/resources/application.properties.template
@@ -1,0 +1,22 @@
+server.port=9093
+
+spring.jackson.deserialization.fail-on-unknown-properties=false
+
+spring.datasource.url=jdbc:h2:mem:darts
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=
+spring.datasource.password=
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.h2.console.enabled=true
+spring.h2.console.settings.web-allow-others=true
+spring.jpa.hibernate.ddl-auto=create-drop
+
+rabbitmq.hostname=
+rabbitmq.username=
+rabbitmq.password=
+rabbitmq.virtualHost=
+rabbitmq.exchange=
+rabbitmq.queue.dialog=
+rabbitmq.routingKey.dialog=
+rabbitmq.queue.game=
+rabbitmq.routingKey.game=
