--- src/main/java/com/kepler/alicedarts/AliceDartsApplication.java
@@ -1,11 +1,0 @@
-package com.kepler.alicedarts;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AliceDartsApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(AliceDartsApplication.class, args);
-    }
-}
