--- src/main/java/com/kepler/alicedarts/model/exchange/request/Account_linking.java
@@ -1,8 +1,0 @@
-package com.kepler.alicedarts.model.exchange.request;
-
-import lombok.Data;
-
-@Data
-public class Account_linking {
-
-}
