--- src/main/java/com/kepler/alicedarts/model/exchange/request/Application.java
@@ -1,8 +1,0 @@
-package com.kepler.alicedarts.model.exchange.request;
-
-import lombok.Data;
-
-@Data
-public class Application {
-    private String application_id;
-}
