--- src/main/java/com/kepler/alicedarts/model/exchange/request/Entities.java
@@ -1,10 +1,0 @@
-package com.kepler.alicedarts.model.exchange.request;
-
-import lombok.Data;
-
-@Data
-public class Entities {
-    private String type;
-    private Tokens tokens;
-    private Object value;
-}
