--- src/main/java/com/kepler/alicedarts/model/exchange/request/InputData.java
@@ -1,12 +1,0 @@
-package com.kepler.alicedarts.model.exchange.request;
-
-import lombok.Data;
-
-@Data
-public class InputData {
-    private Meta meta;
-    private Session session;
-    private Request request;
-    private String version;
-}
-
