--- src/main/java/com/kepler/alicedarts/model/exchange/request/Interfaces.java
@@ -1,10 +1,0 @@
-package com.kepler.alicedarts.model.exchange.request;
-
-import lombok.Data;
-
-@Data
-public class Interfaces {
-    private Screen screen;
-    private Payments payments;
-    private Account_linking account_linking;
-}
