--- src/main/java/com/kepler/alicedarts/model/exchange/request/Nlu.java
@@ -1,10 +1,0 @@
-package com.kepler.alicedarts.model.exchange.request;
-
-import lombok.Data;
-
-@Data
-public class Nlu {
-    private String[] tokens;
-    private Entities[] entities;
-    private Intents intents;
-}
