--- src/main/java/com/kepler/alicedarts/model/exchange/request/Payments.java
@@ -1,8 +1,0 @@
-package com.kepler.alicedarts.model.exchange.request;
-
-import lombok.Data;
-
-@Data
-public class Payments {
-
-}
