--- src/main/java/com/kepler/alicedarts/model/exchange/request/Tokens.java
@@ -1,9 +1,0 @@
-package com.kepler.alicedarts.model.exchange.request;
-
-import lombok.Data;
-
-@Data
-public class Tokens {
-    private Integer start;
-    private Integer end;
-}
