--- src/main/java/com/kepler/alicedarts/model/exchange/request/User.java
@@ -1,8 +1,0 @@
-package com.kepler.alicedarts.model.exchange.request;
-
-import lombok.Data;
-
-@Data
-public class User {
-    private String user_id;
-}
