--- src/main/java/com/kepler/alicedarts/service/PointsCalculator.java
@@ -1,115 +1,0 @@
-package com.kepler.alicedarts.service;
-
-import com.kepler.alicedarts.model.exchange.request.Nlu;
-import lombok.extern.java.Log;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.ListIterator;
-
-@Log
-public class PointsCalculator {
-    public static HashMap<String, String> getPreviousCurrentNextValues(ListIterator<String> wordsIterator) {
-        /*
-        Get previous, current, next values relatively current word
-         */
-        HashMap<String, String> values = new HashMap<>();
-        String curWord;
-        String prevWord;
-        String nextWord;
-        if(wordsIterator.hasPrevious()) {
-            prevWord = wordsIterator.previous();
-            wordsIterator.next();
-        } else {
-            prevWord = null;
-        }
-        curWord = wordsIterator.next();
-        if(wordsIterator.hasNext()) {
-            nextWord = wordsIterator.next();
-            wordsIterator.previous();
-        } else {
-            nextWord = null;
-        }
-        values.put("current", curWord);
-        values.put("previous", prevWord);
-        values.put("next", nextWord);
-        return values;
-    }
-
-    private static int bullOrHalfBull(String word) {
-        /*
-        Check, player hit bull or half-bull
-         */
-        if(word.equals("полубуль")) {
-            return 25;
-        }
-        if(word.equals("буль")) {
-            return 50;
-        }
-        return 0;
-    }
-
-    private static int multipliedSector(String multiplier, int sector) {
-        /*
-        Check, player hit sector with multiplier
-         */
-        if(multiplier != null) {
-            if (multiplier.equals("удвоение")) {
-                return sector * 2;
-            }
-            if (multiplier.equals("утроение")) {
-                return sector * 3;
-            }
-        }
-        return 0;
-    }
-
-    private static int multipliedBullOrHalfBull(int multiplier, String sector) {
-        /*
-        Check, how many times' player hit bull or half-bull
-         */
-        if(sector != null) {
-            if (sector.equals("буля")) {
-                return multiplier * 50;
-            }
-            if (sector.equals("полубуля")) {
-                return multiplier * 25;
-            }
-        }
-        return 0;
-    }
-
-    private static int getMultipliedValueIfExist(HashMap<String, String> values, int curNumber) {
-        /*
-        Calculate, where hit player
-         */
-        return bullOrHalfBull(values.get("current")) +
-                multipliedSector(values.get("previous"), curNumber) +
-                multipliedBullOrHalfBull(curNumber, values.get("next"));
-    }
-
-    public static Integer calculate(Nlu nlu) {
-        int result = 0;
-        List<String> words = new ArrayList<>(List.of(nlu.getTokens()));
-        ListIterator<String> wordsIterator = words.listIterator();
-        HashMap<String, String> values;
-        while(wordsIterator.hasNext()) {
-            values = getPreviousCurrentNextValues(wordsIterator); // try to get previous and next values
-
-            int curNumber = 0;
-            try {
-                curNumber = Integer.parseInt(values.get("current")); // convert string word into int value if it is possible
-            } catch (NumberFormatException ignored) {}
-
-            int multipliedValue = getMultipliedValueIfExist(values, curNumber);
-            // if player not hit any multiplier sector or bull or half-bull, player get scopes for single sector
-            if(multipliedValue == 0) {
-                result += curNumber;
-            } else {
-                result += multipliedValue;
-            }
-
-        }
-        return result;
-    }
-}
