--- src/main/resources/application.properties.template
@@ -1,18 +1,0 @@
-server.port=9092
-
-spring.jackson.deserialization.fail-on-unknown-properties=false
-
-spring.datasource.url=jdbc:h2:mem:darts
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=
-spring.datasource.password=
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-spring.h2.console.enabled=true
-spring.h2.console.settings.web-allow-others=true
-spring.jpa.hibernate.ddl-auto=create-drop
-
-security.require-ssl=true
-server.ssl.key-store-type=PKCS12
-server.ssl.key-store=classpath:keystore.p12
-server.ssl.key-store-password=
-server.ssl.key-alias=
