--- src/test/java/com/kepler/alicedarts/AliceDartsApplicationTests.java
@@ -1,13 +1,0 @@
-package com.kepler.alicedarts;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AliceDartsApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
